fixes when compiler doesn't support PRId64/PRIu64 and some warning fixes

+++ src/eez/flow/lvgl_api.cpp
@@ -297,12 +297,12 @@ bool compareRollerOptions(lv_roller_t *roller, const char *new_val, const char *
     auto n = strlen(new_val);
 
 #if LVGL_VERSION_MAJOR >= 9
-    int numPages = roller->inf_page_cnt;
+    size_t numPages = roller->inf_page_cnt;
 #else
-    int numPages = LV_ROLLER_INF_PAGES;
+    size_t numPages = LV_ROLLER_INF_PAGES;
 #endif
 
-    for (int i = 0; i < numPages * (n + 1); i += n + 1) {
+    for (size_t i = 0; i < numPages * (n + 1); i += n + 1) {
         if (strncmp(new_val, cur_val + i, n) != 0) {
             return true;
         }
